# openapi/rest_mf0226_3.yaml
openapi: 3.0.1
info:
  title: REST MF0226_3
  description: Documentacion REST para la actividad fin de modulo MF0226_3
  version: 1.0.0
servers:
  - url: 'http://localhost:8080/api/'
    description: Servidor de Spring Boot
    variables: {}
paths:
  '/cursos':
    get:
      summary: Ver todos los cursos
      operationId: getAll
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                type: 'array'
                items:
                  $ref: '#/components/schemas/Curso'
  '/cursos/{codigo}':
    get:
      summary: Obtener un curso por id con todas sus reseñas
      operationId: getById
      parameters:
        - name: codigo
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Devuelve el curso cuyo id es el que se ha pedido junto con todas sus reseñas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Curso'
        '404':
          description: No encontrado
  '/resenias/{codigo}':      
    put:
      summary: Actualizar reseña por id
      operationId: update
      parameters:
        - name: codigo
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Resenia'
      responses:
        '400':
           description: Petición incorrecta
        '200':
          description: Modificado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resenia'
    delete:
      summary: Borrar reseña por id
      operationId: delete
      parameters:
        - name: codigo
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Sin contenido
        '404':
          description: No encontrado
  /resenias:
    post:
      summary: Insertar nueva reseña
      operationId: insert
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Resenia'
      responses:
        '201':
          description: Creado
        '400':
           description: Petición incorrecta
components:
  schemas:
    Resenia:
      type: object
      properties:
        codigo:
          type: integer
          format: int64
        resenia:
          type: string
        fecha:
          type: string
          format: date
        alumno_codigo:
          type: integer
          format: int64
        curso_codigo:
          type: integer
          format: int64
    Curso:
      type: object
      properties:
        codigo:
          type: integer
          format: int64
        nombre:
          type: string
        identificador:
          type: string       
        nHoras:
          type: integer
          format: int64  
        profesor:
          type: object
          properties:
            codigo:
              type: integer
              format: int64
            nombre:
              type: string
            apellidos:
              type: string    
        resenas:
          type: array
          items:
            type: object
            properties:
              codigo:
                type: integer
                format: int64
              resenia:
                type: string
              fecha:
                type: string
                format: date
              alumno_codigo:
                type: integer
                format: int64
              curso_codigo:
                type: integer
                format: int64
    Profesor:
      type: object
      properties:
        codigo:
          type: integer
          format: int64
        nombre:
          type: string
        apellidos:
          type: string       
             